Schedule : add field `isLoggingEnabled` with default `false` to close #7
This allows for an application to externally control whether or not a
given Schedule logs.  It does not allow for removal of logging code by
the tree shaker, but that is a general issue for Dart (no external
configuration) so this library cannot do much until Dart figures that
out.